Create lap no from size of FitData List
- instead of mesg.GetMessageIndex because Fit files from
  Wahoo has not message index

Changes to be committed:
	modified:   src/qmapshack/gis/fit2/CFit2Project.cpp
	modified:   src/qmapshack/gis/trk/CFitData.cpp
	modified:   src/qmapshack/gis/trk/CFitData.h
	modified:   src/qmapshack/gis/trk/CFitDataDialog.h

+++ src/qmapshack/gis/trk/CFitData.cpp
@@ -33,7 +33,7 @@ QList<CFitData::lap_t>& CFitData::getLaps()
   return laps;
 }
 
-quint32 CFitData::getNoOfLaps() {
+quint16 CFitData::getNoOfLaps() {
   return laps.size();
 }
